+++ db/migrate/20180910082930_rm_site_relation.rb
@@ -1,0 +1,6 @@
+class RmSiteRelation < ActiveRecord::Migration[5.2]
+  def change
+    remove_reference :folio_nodes, :site
+    remove_reference :visits, :site
+  end
+end