+++ test/controllers/folio/console/nodes_controller_test.rb
@@ -7,10 +7,6 @@ module Folio
   class Console::NodesControllerTest < Console::BaseControllerTest
     include Engine.routes.url_helpers
 
-    setup do
-      @node = create(:folio_node)
-    end
-
     test 'should get index' do
       get console_nodes_url
       assert_response :success
@@ -22,7 +18,8 @@ module Folio
     end
 
     test 'should get edit' do
-      get edit_console_node_url(@node)
+      node = create(:folio_node)
+      get edit_console_node_url(node)
       assert_response :success
     end
   end